refact: Move HashRepository to its own module
