+++ super-agent/src/opamp/mod.rs
@@ -1,5 +1,6 @@
 pub mod callbacks;
 pub mod client_builder;
+pub mod hash_repository;
 pub mod instance_id;
 pub mod operations;
 pub mod remote_config;
