+++ super-agent/src/super_agent/error.rs
@@ -7,9 +7,9 @@ use crate::agent_type::agent_type_registry::AgentRepositoryError;
 use crate::agent_type::agent_values::AgentValuesError;
 use crate::agent_type::error::AgentTypeError;
 use crate::opamp::client_builder::OpAMPClientBuilderError;
+use crate::opamp::hash_repository::HashRepositoryError;
 use crate::opamp::instance_id;
 use crate::opamp::remote_config::RemoteConfigError;
-use crate::opamp::remote_config_hash::HashRepositoryError;
 use crate::sub_agent::effective_agents_assembler::EffectiveAgentsAssemblerError;
 use crate::sub_agent::error::{SubAgentBuilderError, SubAgentCollectionError, SubAgentError};
 use crate::sub_agent::persister::config_persister::PersistError;
